Move the android package to the common module
