+++ common/src/main/java/com/alexstyl/android/AlarmManagerCompat.java
@@ -4,8 +4,6 @@ import android.app.AlarmManager;
 import android.app.PendingIntent;
 import android.content.Context;
 
-import com.alexstyl.specialdates.util.Utils;
-
 public final class AlarmManagerCompat {
     private final AlarmManager alarmManager;
 
@@ -19,7 +17,7 @@ public final class AlarmManagerCompat {
     }
 
     public void setExact(int type, long triggerAtmillis, PendingIntent operation) {
-        if (Utils.hasKitKat()) {
+        if (Version.hasKitKat()) {
             alarmManager.setExact(type, triggerAtmillis, operation);
         } else {
             alarmManager.set(AlarmManager.RTC, triggerAtmillis, operation);
