+++ common/src/main/java/com/alexstyl/android/AndroidColorResources.java
