+++ common/src/main/java/com/alexstyl/android/AndroidDimensionResources.java
