+++ common/src/main/java/com/alexstyl/android/ViewVisibility.java
