+++ mobile/src/main/res/values/attrs.xml
@@ -18,15 +18,7 @@
       format="string" />
   </declare-styleable>
 
-  <declare-styleable name="ForegroundLinearLayout">
-    <attr name="android:foreground" />
-    <attr name="android:foregroundInsidePadding" />
-    <attr name="android:foregroundGravity" />
-  </declare-styleable>
 
-  <declare-styleable name="ForegroundImageView">
-    <attr name="android:foreground" />
-  </declare-styleable>
 
   <declare-styleable name="TitledTextView">
     <attr name="eventTitleColor"
